assume-test/requirements_assume.txt
argcomplete==3.6.3
assume-framework==0.5.5
colorama==0.4.6
dill==0.4.0
fsspec==2025.10.0
greenlet==3.2.4
highspy==1.12.0
mango-agents==2.1.5
nest-asyncio==1.6.0
networkx==3.6
numpy==2.3.5
packaging==25.0
paho-mqtt==2.1.0
pandas==2.3.3
ply==3.11
protobuf==5.27.2
psycopg2-binary==2.9.11
pyomo==6.9.5
python-dateutil==2.9.0.post0
pytz==2025.2
PyYAML==6.0.3
pyyaml-include==2.2
setuptools==80.9.0
six==1.17.0
SQLAlchemy==2.0.44
tqdm==4.67.1
typing_extensions==4.15.0
tzdata==2025.2
wheel==0.45.1
xarray==2025.6.1
